Extracted AffinityPropagation inner classes.
