+++ src/main/java/ml/grafos/okapi/clustering/ap/MasterComputation.java
@@ -1,0 +1,34 @@
+/**
+ * Copyright 2014 Grafos.ml
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ml.grafos.okapi.clustering.ap;
+
+import org.apache.giraph.master.DefaultMasterCompute;
+
+/**
+* Affinity Propagation's MasterCompute.
+*
+* @author Toni Penya-Alba <tonipenya@iiia.csic.es>
+* @author Marc Pujol-Gonzalez <mpujol@iiia.csic.es>
+*/
+public class MasterComputation extends DefaultMasterCompute {
+
+  @Override
+  public void initialize() throws InstantiationException, IllegalAccessException {
+    super.initialize();
+    registerPersistentAggregator("exemplars", ExemplarAggregator.class);
+  }
+
+}
